+++ packages/fluent-pouchdb/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.5.20
 
+## 0.9.31
+
+### Patch Changes
+
+- @goatlab/fluent@0.9.31
+
 ## 0.9.30
 
 ### Patch Changes